Move IstreamEnumeration into the acommon namespace.
